requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
groq==0.5.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
pyjwt==2.8.0
